Move stream-related utilities out of Util and into DaggerStreams.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=173319213
