+++ java/dagger/internal/codegen/BindsInstanceProcessingStep.java
@@ -20,11 +20,11 @@ import static com.google.common.collect.Iterables.getOnlyElement;
 import static dagger.internal.codegen.ConfigurationAnnotations.getComponentOrSubcomponentAnnotation;
 import static dagger.internal.codegen.ConfigurationAnnotations.getModuleAnnotation;
 import static dagger.internal.codegen.DaggerElements.isAnyAnnotationPresent;
+import static dagger.internal.codegen.DaggerStreams.toImmutableSet;
 import static dagger.internal.codegen.ErrorMessages.BINDS_INSTANCE_IN_INVALID_COMPONENT;
 import static dagger.internal.codegen.ErrorMessages.BINDS_INSTANCE_IN_MODULE;
 import static dagger.internal.codegen.ErrorMessages.BINDS_INSTANCE_ONE_PARAMETER;
 import static dagger.internal.codegen.MoreAnnotationMirrors.simpleName;
-import static dagger.internal.codegen.Util.toImmutableSet;
 import static javax.lang.model.element.Modifier.ABSTRACT;
 
 import com.google.auto.common.BasicAnnotationProcessor.ProcessingStep;
